+++ tests/test_ranges.py
@@ -1019,6 +1019,60 @@ class TestFiniteDateRange:
 
 
 class TestFiniteDatetimeRange:
+    @pytest.mark.parametrize(
+        "r1, r2, expected",
+        [
+            [
+                ranges.FiniteDatetimeRange(
+                    start=datetime.datetime(2000, 1, 1),
+                    end=datetime.datetime(2000, 1, 3),
+                ),
+                ranges.FiniteDatetimeRange(
+                    start=datetime.datetime(2000, 1, 2),
+                    end=datetime.datetime(2000, 1, 4),
+                ),
+                True,
+            ],
+            [
+                ranges.FiniteDatetimeRange(
+                    start=datetime.datetime(2000, 1, 2),
+                    end=datetime.datetime(2000, 1, 4),
+                ),
+                ranges.FiniteDatetimeRange(
+                    start=datetime.datetime(2000, 1, 1),
+                    end=datetime.datetime(2000, 1, 3),
+                ),
+                False,
+            ],
+            [
+                ranges.FiniteDatetimeRange(
+                    start=datetime.datetime(2000, 1, 1),
+                    end=datetime.datetime(2000, 1, 3),
+                ),
+                ranges.FiniteDatetimeRange(
+                    start=datetime.datetime(2000, 1, 1),
+                    end=datetime.datetime(2000, 1, 5),
+                ),
+                # False, since only `start` is considered.
+                False,
+            ],
+            [
+                ranges.FiniteDatetimeRange(
+                    start=datetime.datetime(2000, 1, 1),
+                    end=datetime.datetime(2000, 1, 3),
+                ),
+                ranges.Range(
+                    start=None,
+                    end=datetime.datetime(3000, 1, 1),
+                    boundaries=ranges.RangeBoundaries.EXCLUSIVE_EXCLUSIVE,
+                ),
+                False,
+            ],
+        ],
+    )
+    def test__lt__(self, r1, r2, expected):
+        assert (r1 < r2) is expected
+
     class TestUnion:
         def test_union_of_touching_ranges(self):
             range = ranges.FiniteDatetimeRange(
@@ -1030,11 +1084,13 @@ class TestFiniteDatetimeRange:
                 end=datetime.datetime(2000, 1, 3),
             )
 
-            union = range | other
-
-            assert union == ranges.FiniteDatetimeRange(
-                start=datetime.datetime(2000, 1, 1),
-                end=datetime.datetime(2000, 1, 3),
+            assert (
+                range | other
+                == other | range
+                == ranges.FiniteDatetimeRange(
+                    start=datetime.datetime(2000, 1, 1),
+                    end=datetime.datetime(2000, 1, 3),
+                )
             )
 
         def test_union_of_disjoint_ranges(self):
@@ -1047,7 +1103,7 @@ class TestFiniteDatetimeRange:
                 end=datetime.datetime(2020, 1, 2),
             )
 
-            assert range | other is None
+            assert (range | other is None) and (other | range is None)
 
         def test_union_of_overlapping_ranges(self):
             range = ranges.FiniteDatetimeRange(
@@ -1059,11 +1115,57 @@ class TestFiniteDatetimeRange:
                 end=datetime.datetime(2000, 1, 4),
             )
 
-            union = range | other
+            assert (
+                range | other
+                == other | range
+                == ranges.FiniteDatetimeRange(
+                    start=datetime.datetime(2000, 1, 1),
+                    end=datetime.datetime(2000, 1, 4),
+                )
+            )
 
-            assert union == ranges.FiniteDatetimeRange(
+    class TestIntersection:
+        def test_intersection_of_touching_ranges(self):
+            range = ranges.FiniteDatetimeRange(
                 start=datetime.datetime(2000, 1, 1),
+                end=datetime.datetime(2000, 1, 2),
+            )
+            other = ranges.FiniteDatetimeRange(
+                start=datetime.datetime(2000, 1, 2),
+                end=datetime.datetime(2000, 1, 3),
+            )
+
+            assert (range & other is None) and (other & range is None)
+
+        def test_intersection_of_disjoint_ranges(self):
+            range = ranges.FiniteDateRange(
+                start=datetime.datetime(2000, 1, 1),
+                end=datetime.datetime(2000, 1, 2),
+            )
+            other = ranges.FiniteDatetimeRange(
+                start=datetime.datetime(2020, 1, 1),
+                end=datetime.datetime(2020, 1, 2),
+            )
+
+            assert (range & other is None) and (other & range is None)
+
+        def test_intersection_of_overlapping_ranges(self):
+            range = ranges.FiniteDatetimeRange(
+                start=datetime.datetime(2000, 1, 1),
+                end=datetime.datetime(2000, 1, 3),
+            )
+            other = ranges.FiniteDatetimeRange(
+                start=datetime.datetime(2000, 1, 2),
                 end=datetime.datetime(2000, 1, 4),
+            )
+
+            assert (
+                range & other
+                == other & range
+                == ranges.FiniteDatetimeRange(
+                    start=datetime.datetime(2000, 1, 2),
+                    end=datetime.datetime(2000, 1, 3),
+                )
             )
 
     class TestLocalize: